Moved calendar codes to ImageActivity
