+++ app/src/main/java/com/illegaldisease/banreco/activities/ImageActivity.kt
@@ -6,14 +6,19 @@ import android.support.design.widget.FloatingActionButton
 import android.widget.ImageView
 import com.illegaldisease.banreco.R
 import com.illegaldisease.banreco.databaserelated.EventHandler
+import com.wdullaer.materialdatetimepicker.date.DatePickerDialog
+import com.wdullaer.materialdatetimepicker.time.TimePickerDialog
+import java.util.*
 
-class ImageActivity : AppCompatActivity() {
+class ImageActivity : AppCompatActivity(), TimePickerDialog.OnTimeSetListener,DatePickerDialog.OnDateSetListener {
 
     private lateinit var imageView : ImageView
     private lateinit var trashButton : FloatingActionButton
     private lateinit var setDateButton : FloatingActionButton
     private lateinit var setTimeButton : FloatingActionButton
     private lateinit var doneButton : FloatingActionButton
+
+    private var lastEventDate : Calendar = GregorianCalendar.getInstance(TimeZone.getDefault()) //Don't forget to re-initialize
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -43,4 +48,35 @@ class ImageActivity : AppCompatActivity() {
         setTimeButton.setOnClickListener { /*TODO: Open TimeDialogPicker here */}
         doneButton.setOnClickListener { /*TODO: Add to database here */ }
     }
+    private fun pickTime(){
+        val tpd = TimePickerDialog.newInstance(
+                this@ImageActivity,
+                true
+        )
+        tpd.show(fragmentManager,"TimePicker")
+        tpd.version = TimePickerDialog.Version.VERSION_2
+    }
+    private fun pickDate(){
+        val now = Calendar.getInstance()
+        val dpd = DatePickerDialog.newInstance(
+                this@ImageActivity,
+                now.get(Calendar.YEAR),
+                now.get(Calendar.MONTH),
+                now.get(Calendar.DAY_OF_MONTH)
+        )
+        dpd.show(fragmentManager, "DatePicker")
+        dpd.version = DatePickerDialog.Version.VERSION_2
+    }
+    override fun onDateSet(view: DatePickerDialog?, year: Int, monthOfYear: Int, dayOfMonth: Int) {
+        lastEventDate.set(Calendar.MONTH, monthOfYear)
+        lastEventDate.set(Calendar.YEAR, year)
+        lastEventDate.set(Calendar.DAY_OF_MONTH, dayOfMonth)
+        pickTime()
+    }
+    override fun onTimeSet(view: TimePickerDialog?, hourOfDay: Int, minute: Int, second: Int) {
+        lastEventDate.set(Calendar.HOUR_OF_DAY, hourOfDay)
+        lastEventDate.set(Calendar.MINUTE, minute)
+        lastEventDate.set(Calendar.SECOND, second)
+    }
+
 }
